v1.0
Allows 10 different decryption methods from the command line, accepting either a string or a file as a source and allows saving to screen, a default file path or a user defined file. Useful for cybersecurity personnel, without the need to rely on external resources, thereby exposing potentially sensitive data to entities outside the network.